Seems to only be there partially

+++ scripts/patch-dotnet.ps1
@@ -10,9 +10,12 @@ if (Test-Path $dotnetSdk) {
   foreach ($version in $versions) {
     $root = Join-Path $version "Sdks/Microsoft.NET.Sdk.WindowsDesktop/targets"
     if (Test-Path $root) {
-      dir $root
-      Move-Item (Join-Path $root Microsoft.WinFx.props) (Join-Path $root Microsoft.WinFX.props)
-      Move-Item (Join-Path $root Microsoft.WinFx.targets) (Join-Path $root Microsoft.WinFX.targets)
+      if (Test-Path (Join-Path $root Microsoft.WinFx.props)) {
+        Move-Item (Join-Path $root Microsoft.WinFx.props) (Join-Path $root Microsoft.WinFX.props)
+      }
+      if (Test-Path (Join-Path $root Microsoft.WinFx.targets)) {
+        Move-Item (Join-Path $root Microsoft.WinFx.targets) (Join-Path $root Microsoft.WinFX.targets)
+      }
     }
   }
 } else {
